// apps/desktop/src/global.d.ts
export interface PrometheusReadDirEntry {
  name: string
  path: string
  isDirectory: boolean
}

export interface PrometheusReadDirResult {
  entries: PrometheusReadDirEntry[]
  error?: string
}